chore: move format_delta to better_updated
I would like to import it from the scraping script without running
the main function in dashboard.py.
Arguably, that function belongs in a basic utility module anyway.
